(#14) Rename is_inside, is_outside in former Corridor
to Walls::contains() which is former is_outside.
In the past Corridor's outside meant walls, Corridor's inside meant the space between walls. From now on, the structure is named as Walls. Instead of Corridor's is_outside: Walls::contains is beeing used, which I beleieve is more intuitive. The notion of Corridor is not used any more.